First graphical accounting of IPv6.
There are two small changes to "http.c" that corrects
the format string and that reuses the exact length of
the structure in use.

The main changes, however, deal with detection, reporting,
and accounting of IPv6 traffic over the chosen interface.

The present code is only able to match the exact
IPv6 address of the interface to the packet, be it
inbound or outbound. No possibility to set the
netmask length.

The accounting only adds sizes of the packets,
it never goes on to deal with fine grained information.
You will see a conditional in 'acct.c' with a comment
stating that IPv6 proceeds no further.