Improve graphs page display on mobile devices.
This has little to no effect on a desktop browser, but on a mobile browser, it
will automatically fit the page to the browser window, without the user having
to manually zoom or scroll left or right to see everything. This results in the
graphs being arranged in a single column, rather then a 2x2 grid, which is also
what happens on my desktop when I make the browser window narrower.